Introduce dedicated header for window-related enum types
