+++ src/openrct2-ui/interface/FileBrowser.h
@@ -16,13 +16,14 @@
 #include <openrct2/localisation/StringIdType.h>
 #include <string>
 
-enum class LoadSaveAction : uint8_t;
-enum class LoadSaveType : uint8_t;
-enum class ModalResult : int8_t;
 struct TrackDesign;
 
 namespace OpenRCT2
 {
+    enum class LoadSaveAction : uint8_t;
+    enum class LoadSaveType : uint8_t;
+    enum class ModalResult : int8_t;
+
     struct WindowBase;
 }
