+++ src/openrct2/interface/Window.h
@@ -11,15 +11,18 @@
 
 #include "../Identifiers.h"
 #include "../core/EnumUtils.hpp"
+#include "WindowTypes.h"
 
 #include <functional>
 #include <memory>
 
-enum class CloseWindowModifier : uint8_t;
-enum class WindowClass : uint8_t;
-
 struct CoordsXYZ;
 struct ZoomLevel;
+
+// Cannot be expressed using a constexpr.
+// NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
+#define VALIDATE_GLOBAL_WIDX(wc, widx)                                                                                         \
+    static_assert(widx == wc##__##widx, "Global WIDX of " #widx " doesn't match actual value.")
 
 namespace OpenRCT2::Drawing
 {
@@ -31,155 +34,13 @@ namespace OpenRCT2
     using WindowNumber = int16_t;
     using WidgetIndex = uint16_t;
 
+    struct Viewport;
     struct WindowBase;
     struct WindowCloseModifier;
 
-    struct WindowIdentifier
-    {
-        WindowClass classification;
-        WindowNumber number;
-    };
-
-    struct WidgetIdentifier
-    {
-        WindowIdentifier window;
-        WidgetIndex widgetIndex;
-    };
-
     extern WindowCloseModifier gLastCloseModifier;
-
-    struct WindowCloseModifier
-    {
-        WindowIdentifier window;
-        CloseWindowModifier modifier;
-    };
-
-    struct Viewport;
-
-    enum class WindowView : uint8_t
-    {
-        parkAwards,
-        parkRating,
-        parkObjective,
-        parkGuests,
-        financesResearch,
-        rideResearch,
-        mazeConstruction,
-        networkPassword,
-        editorBottomToolbar,
-        changelog,
-        newVersionInfo,
-        financeMarketing,
-        contributors,
-    };
-
-    enum class WindowDetail : uint8_t
-    {
-        banner,
-        newCampaign,
-        demolishRide,
-        refurbishRide,
-        sign,
-        signSmall,
-
-        player,
-
-        vehicle,
-        track,
-
-        null = 255,
-    };
-} // namespace OpenRCT2
-
-// Cannot be expressed using a constexpr.
-// NOLINTNEXTLINE(cppcoreguidelines-macro-usage)
-#define VALIDATE_GLOBAL_WIDX(wc, widx)                                                                                         \
-    static_assert(widx == wc##__##widx, "Global WIDX of " #widx " doesn't match actual value.")
-
-enum class PromptMode : uint8_t
-{
-    saveBeforeLoad = 0,
-    saveBeforeQuit,
-    saveBeforeQuit2,
-    saveBeforeNewGame,
-    quit
-};
-
-enum BTM_TOOLBAR_DIRTY_FLAGS
-{
-    BTM_TB_DIRTY_FLAG_MONEY = (1 << 0),
-    BTM_TB_DIRTY_FLAG_DATE = (1 << 1),
-    BTM_TB_DIRTY_FLAG_PEEP_COUNT = (1 << 2),
-    BTM_TB_DIRTY_FLAG_CLIMATE = (1 << 3),
-    BTM_TB_DIRTY_FLAG_PARK_RATING = (1 << 4)
-};
-
-enum class LoadSaveAction : uint8_t
-{
-    load,
-    save,
-};
-
-enum class LoadSaveType : uint8_t
-{
-    park,
-    landscape,
-    scenario,
-    track,
-    heightmap,
-};
-
-enum class ModalResult : int8_t
-{
-    fail = -1,
-    cancel,
-    ok,
-};
-
-enum class CloseWindowModifier : uint8_t
-{
-    none,
-    shift,
-    control
-};
-
-enum class GuestListFilterType : int32_t
-{
-    guestsOnRide,
-    guestsInQueue,
-    guestsThinkingAboutRide,
-    guestsThinkingX,
-};
-
-enum class Tool
-{
-    arrow = 0,
-    upArrow = 2,
-    upDownArrow = 3,
-    picker = 7,
-    crosshair = 12,
-    pathDown = 17,
-    digDown = 18,
-    waterDown = 19,
-    walkDown = 22,
-    paintDown = 23,
-    entranceDown = 24,
-    bulldozer = 27,
-};
-
-namespace OpenRCT2
-{
-    struct WidgetRef
-    {
-        WindowClass windowClassification;
-        WindowNumber windowNumber;
-        WidgetIndex widgetIndex;
-    };
-
     extern Tool gCurrentToolId;
     extern WidgetRef gCurrentToolWidget;
-
-    using CloseCallback = void (*)();
 
     constexpr int8_t kWindowLimitMin = 4;
     constexpr int8_t kWindowLimitMax = 64;
