+++ src/openrct2/ui/WindowManager.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "../interface/Window.h"
+#include "../Identifiers.h"
 #include "../interface/WindowFlags.h"
 #include "../localisation/StringIdType.h"
 #include "../windows/Intent.h"
