+++ src/openrct2/windows/Intent.h
@@ -10,9 +10,9 @@
 #pragma once
 
 #include "../core/Identifier.hpp"
-#include "../interface/Window.h"
 #include "../interface/WindowClasses.h"
 #include "../interface/WindowFlags.h"
+#include "../interface/WindowTypes.h"
 
 #include <map>
 #include <sfl/static_vector.hpp>
